requirements.txt
requests
beautifulsoup4
pycrypto